[MIPS] This commit includes interpreter loop and Opcode optimizations for MIPS.
There is a performance gain of ~=20% in Caffeine Mark as compared to the portable
interpreter.

The MIPS Opcodes are picked up from vm/mterp/mips directory. The handler size is
increased to 128 bytes as compared to the default 64. This is done to prevent
excessive branching.

Caveats:
1. common_periodicChecks is currently ignored in the interpreter. This did not
cause any problems with dalvik test suite.
2. 6 more opcodes need to be converted to MIPS. Currently their C counterparts are
being used. (See config-mips)
3. The hard floating point support is not yet enabled in MIPS interpreter.
Further work is required before enabling it.

   Signed-off-by: Chris Dearman <chris@mips.com>
   Signed-off-by: Steve Weissinger <stevew@mips.com>
   Signed-off-by: Raghu Gandham <raghu@mips.com>