[MIPS] Fix new style chaining
When the chaining implementation was changed to include an initial
branch as was done for ARM, the chaining code was not modified to
patch the new branch. The net effect was that chaining worked, but
unchaining didn't.

(cherry picked from commit daec078fa6ed415e9ef63cea19b386268b0498f1)

Change-Id: Icd63a137e4e4190aa74e294c214c82c9131bff2c